python/requirements.txt
pika
pybrain
scipy
numpy
pyyaml
boto
tflearn
pathlib
svgwrite